refactor: 회원가입 API 명세 변경에 따른 수정
- 회원가입 기능을 auth 패키지로 이동
