+++ backend/src/main/java/com/woowacourse/momo/auth/controller/AuthController.java
@@ -2,6 +2,7 @@ package com.woowacourse.momo.auth.controller;
 
 import javax.validation.Valid;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,8 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 
-import com.woowacourse.momo.auth.dto.LoginRequest;
-import com.woowacourse.momo.auth.dto.LoginResponse;
+import com.woowacourse.momo.auth.dto.request.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
+import com.woowacourse.momo.auth.dto.response.LoginResponse;
 import com.woowacourse.momo.auth.service.AuthService;
 
 @RequiredArgsConstructor
@@ -26,5 +28,12 @@ public class AuthController {
         LoginResponse response = authService.login(request);
 
         return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/signup")
+    public ResponseEntity<Void> signUp(@RequestBody @Valid SignUpRequest request) {
+        authService.signUp(request);
+
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 }
