+++ backend/src/main/java/com/woowacourse/momo/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.auth.dto;
+package com.woowacourse.momo.auth.dto.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
