+++ backend/src/main/java/com/woowacourse/momo/member/controller/MemberController.java
@@ -1,20 +1,14 @@
 package com.woowacourse.momo.member.controller;
 
-import javax.validation.Valid;
-
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.auth.config.AuthenticationPrincipal;
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
 import com.woowacourse.momo.member.dto.response.MemberResponse;
 import com.woowacourse.momo.member.service.MemberService;
 
@@ -25,21 +19,14 @@ public class MemberController {
 
     private final MemberService memberService;
 
-    @PostMapping
-    public ResponseEntity<Void> signUp(@RequestBody @Valid SignUpRequest request) {
-        memberService.signUp(request);
-
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @GetMapping("/info")
+    @GetMapping
     public ResponseEntity<MemberResponse> find(@AuthenticationPrincipal Long id) {
         MemberResponse response = memberService.findById(id);
 
         return ResponseEntity.ok(response);
     }
 
-    @DeleteMapping()
+    @DeleteMapping
     public ResponseEntity<Void> delete(@AuthenticationPrincipal Long id) {
         memberService.deleteById(id);
