+++ backend/src/main/java/com/woowacourse/momo/member/service/MemberService.java
@@ -7,8 +7,6 @@ import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.member.domain.Member;
 import com.woowacourse.momo.member.domain.MemberRepository;
-import com.woowacourse.momo.member.domain.PasswordEncoder;
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
 import com.woowacourse.momo.member.dto.response.MemberResponse;
 import com.woowacourse.momo.member.exception.NotFoundMemberException;
 
@@ -18,16 +16,6 @@ import com.woowacourse.momo.member.exception.NotFoundMemberException;
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Transactional
-    public Long signUp(SignUpRequest request) {
-        String password = passwordEncoder.encrypt(request.getPassword());
-        Member member = new Member(request.getEmail(), password, request.getName());
-        Member savedMember = memberRepository.save(member);
-
-        return savedMember.getId();
-    }
 
     public MemberResponse findById(Long id) {
         Member member = memberRepository.findById(id)
