+++ backend/src/test/java/com/woowacourse/momo/auth/acceptance/AuthAcceptanceTest.java
@@ -6,7 +6,8 @@ import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import com.woowacourse.momo.auth.dto.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
 import com.woowacourse.momo.common.acceptance.AcceptanceTest;
 import com.woowacourse.momo.common.acceptance.RestAssuredConvenienceMethod;
 
@@ -15,6 +16,14 @@ public class AuthAcceptanceTest extends AcceptanceTest {
 
     private static final String EMAIL = "woowa@woowa.com";
     private static final String PASSWORD = "1q2w3e4R!";
+
+    @Test
+    void 회원_가입_테스트() {
+        SignUpRequest request = new SignUpRequest(EMAIL, PASSWORD, "모모");
+
+        RestAssuredConvenienceMethod.postRequest(request, "/api/auth/signup")
+                .statusCode(HttpStatus.CREATED.value());
+    }
 
     @Test
     public void 로그인_테스트() {
