+++ backend/src/test/java/com/woowacourse/momo/auth/controller/AuthControllerTest.java
@@ -6,7 +6,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,14 +17,18 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import com.woowacourse.momo.auth.dto.LoginRequest;
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
-import com.woowacourse.momo.member.service.MemberService;
+import com.woowacourse.momo.auth.dto.request.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
+import com.woowacourse.momo.auth.service.AuthService;
 
 @Transactional
 @AutoConfigureMockMvc
 @SpringBootTest
 class AuthControllerTest {
+
+    private static final String EMAIL = "woowa@woowa.com";
+    private static final String PASSWORD = "wooteco1!";
+    private static final String NAME = "모모";
 
     @Autowired
     private MockMvc mockMvc;
@@ -34,18 +37,96 @@ class AuthControllerTest {
     private ObjectMapper objectMapper;
 
     @Autowired
-    private MemberService memberService;
+    private AuthService authService;
 
-    @BeforeEach
-    void setUp() {
-        SignUpRequest request = new SignUpRequest("woowa@woowa.com", "Woowa1!", "모모");
-        memberService.signUp(request);
+    @DisplayName("정상적으로 회원가입이 되는 경우를 테스트한다.")
+    @Test
+    void signUp() throws Exception {
+        SignUpRequest request = new SignUpRequest(EMAIL, PASSWORD, NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .content(objectMapper.writeValueAsString(request))
+        ).andExpect(status().isCreated());
+    }
+
+    @DisplayName("잘못된 이메일 형식으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithInvalidEmailPattern() throws Exception {
+        SignUpRequest request = new SignUpRequest("woowa", PASSWORD, NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("잘못된 이메일 형식입니다.")));
+    }
+
+    @DisplayName("비어있는 이메일 값으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithBlankEmail() throws Exception {
+        SignUpRequest request = new SignUpRequest("", PASSWORD, NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("이메일은 빈 값일 수 없습니다.")));
+    }
+
+    @DisplayName("비어있는 비밀번호 값으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithBlankPassword() throws Exception {
+        SignUpRequest request = new SignUpRequest(EMAIL, "", NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("패스워드는 빈 값일 수 없습니다.")));
+    }
+
+    @DisplayName("잘못된 비밀번호 패턴으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithInvalidPasswordPattern() throws Exception {
+        SignUpRequest request = new SignUpRequest(EMAIL, "woowa", NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("패스워드는 소문자와 하나 이상의 숫자, 문자를 갖고 있어야 합니다.")));
+    }
+
+    @DisplayName("잘못된 비밀번호 패턴으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithInvalidPasswordPattern1() throws Exception {
+        SignUpRequest request = new SignUpRequest(EMAIL, "woowacourse1", NAME);
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("패스워드는 소문자와 하나 이상의 숫자, 문자를 갖고 있어야 합니다.")));
+    }
+
+    @DisplayName("비어있는 이름 값으로 회원가입시 400코드가 반환된다.")
+    @Test
+    void signUpWithBlankName() throws Exception {
+        SignUpRequest request = new SignUpRequest("woowa@woowa.com", PASSWORD, "");
+
+        mockMvc.perform(post("/api/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .content(objectMapper.writeValueAsString(request))
+                ).andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$", containsString("이름은 빈 값일 수 없습니다.")));
     }
 
     @DisplayName("정상적으로 로그인될 시 토큰이 발급된다.")
     @Test
     void login() throws Exception {
-        LoginRequest request = new LoginRequest("woowa@woowa.com", "Woowa1!");
+        createNewMember(EMAIL, PASSWORD, NAME);
+        LoginRequest request = new LoginRequest(EMAIL, PASSWORD);
 
         mockMvc.perform(post("/api/auth/login")
                         .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -57,7 +138,7 @@ class AuthControllerTest {
     @DisplayName("잘못된 이메일 형식으로 로그인시 400코드가 반환된다.")
     @Test
     void loginWithInvalidEmailPattern() throws Exception {
-        LoginRequest request = new LoginRequest("woowa", "Woowa1!");
+        LoginRequest request = new LoginRequest("woowa", PASSWORD);
 
         mockMvc.perform(post("/api/auth/login")
                         .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -69,7 +150,8 @@ class AuthControllerTest {
     @DisplayName("비어있는 이메일 값으로 로그인시 400코드가 반환된다.")
     @Test
     void loginWithBlankEmail() throws Exception {
-        LoginRequest request = new LoginRequest("", "Woowa1!");
+        createNewMember(EMAIL, PASSWORD, NAME);
+        LoginRequest request = new LoginRequest("", PASSWORD);
 
         mockMvc.perform(post("/api/auth/login")
                         .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -81,6 +163,7 @@ class AuthControllerTest {
     @DisplayName("비어있는 비밀번호 형식으로 로그인시 400코드가 반환된다.")
     @Test
     void loginWithBlankPassword() throws Exception {
+        createNewMember(EMAIL, PASSWORD, NAME);
         LoginRequest request = new LoginRequest("woowa@woowa.com", "");
 
         mockMvc.perform(post("/api/auth/login")
@@ -88,5 +171,10 @@ class AuthControllerTest {
                         .content(objectMapper.writeValueAsString(request))
                 ).andExpect(status().isBadRequest())
                 .andExpect(jsonPath("$", containsString("패스워드는 빈 값일 수 없습니다.")));
+    }
+
+    void createNewMember(String email, String password, String name) {
+        SignUpRequest request = new SignUpRequest(email, password, name);
+        authService.signUp(request);
     }
 }
