+++ backend/src/test/java/com/woowacourse/momo/auth/service/AuthServiceTest.java
@@ -1,19 +1,18 @@
 package com.woowacourse.momo.auth.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.woowacourse.momo.auth.dto.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
 import com.woowacourse.momo.auth.exception.AuthFailException;
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
-import com.woowacourse.momo.member.service.MemberService;
 
 @Transactional
 @SpringBootTest
@@ -21,22 +20,24 @@ class AuthServiceTest {
 
     private static final String EMAIL = "woowa@woowa.com";
     private static final String PASSWORD = "wooteco1!";
+    private static final String NAME = "모모";
 
     @Autowired
     private AuthService authService;
 
-    @Autowired
-    private MemberService memberService;
+    @DisplayName("회원 가입을 한다")
+    @Test
+    void signUp() {
+        SignUpRequest request = new SignUpRequest(EMAIL, PASSWORD, NAME);
+        Long id = authService.signUp(request);
 
-    @BeforeEach
-    void setUp() {
-        SignUpRequest request = new SignUpRequest(EMAIL, PASSWORD, "모모");
-        memberService.signUp(request);
+        assertThat(id).isNotNull();
     }
 
     @DisplayName("로그인을 성공한다")
     @Test
     void login() {
+        createMember(EMAIL, PASSWORD, NAME);
         LoginRequest request = new LoginRequest(EMAIL, PASSWORD);
 
         assertDoesNotThrow(() -> authService.login(request));
@@ -45,9 +46,15 @@ class AuthServiceTest {
     @DisplayName("로그인에 실패한다")
     @Test
     void loginFail() {
+        createMember(EMAIL, PASSWORD, NAME);
         LoginRequest request = new LoginRequest(EMAIL, "wrongPassword");
 
         assertThatThrownBy(() -> authService.login(request))
                 .isInstanceOf(AuthFailException.class);
+    }
+
+    void createMember(String email, String password, String name) {
+        SignUpRequest request = new SignUpRequest(email, password, name);
+        authService.signUp(request);
     }
 }
