+++ backend/src/test/java/com/woowacourse/momo/common/acceptance/Fixture.java
@@ -1,7 +1,7 @@
 package com.woowacourse.momo.common.acceptance;
 
-import com.woowacourse.momo.auth.dto.LoginRequest;
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
+import com.woowacourse.momo.auth.dto.request.LoginRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class Fixture {
@@ -9,7 +9,7 @@ public class Fixture {
     public static void 회원_가입(String email, String password, String name) {
         SignUpRequest request = new SignUpRequest(email, password, name);
 
-        RestAssuredConvenienceMethod.postRequest(request, "/api/members");
+        RestAssuredConvenienceMethod.postRequest(request, "/api/auth/signup");
     }
 
     public static String 로그인(String email, String password) {
