+++ backend/src/test/java/com/woowacourse/momo/member/domain/MemberRepositoryTest.java
@@ -20,7 +20,6 @@ class MemberRepositoryTest {
     @Autowired
     private EntityManager entityManager;
 
-
     @DisplayName("회원을 저장한다.")
     @Test
     void save() {
