+++ backend/src/test/java/com/woowacourse/momo/member/service/MemberServiceTest.java
@@ -9,7 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.woowacourse.momo.member.dto.request.SignUpRequest;
+import com.woowacourse.momo.auth.dto.request.SignUpRequest;
+import com.woowacourse.momo.auth.service.AuthService;
 import com.woowacourse.momo.member.dto.response.MemberResponse;
 import com.woowacourse.momo.member.exception.NotFoundMemberException;
 
@@ -20,20 +21,14 @@ class MemberServiceTest {
     @Autowired
     private MemberService memberService;
 
-    @DisplayName("회원 가입을 한다")
-    @Test
-    void signUp() {
-        SignUpRequest request = new SignUpRequest("woowa@woowa.com", "wooteco1!", "모모");
-        Long id = memberService.signUp(request);
-
-        assertThat(id).isNotNull();
-    }
+    @Autowired
+    private AuthService authService;
 
     @DisplayName("회원 정보를 조회한다.")
     @Test
     void findById() {
         SignUpRequest request = new SignUpRequest("woowa@woowa.com", "wooteco1!", "모모");
-        Long memberId = memberService.signUp(request);
+        Long memberId = authService.signUp(request);
 
         MemberResponse response = memberService.findById(memberId);
 
@@ -46,7 +41,7 @@ class MemberServiceTest {
     @Test
     void delete() {
         SignUpRequest request = new SignUpRequest("woowa@woowa.com", "wooteco1!", "모모");
-        Long memberId = memberService.signUp(request);
+        Long memberId = authService.signUp(request);
 
         memberService.deleteById(memberId);
